requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# MongoDB
motor==3.6.0
pymongo==4.9.2
beanie==1.26.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation and settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis for caching and rate limiting
redis==5.0.1
aioredis==2.0.1

# HTTP client for social auth
httpx==0.26.0

# WebSocket
websockets==12.0

# Image handling
Pillow==10.2.0
python-magic==0.4.27

# AWS S3 / DigitalOcean Spaces
boto3==1.34.0

# Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
